+++ js/photos.js
@@ -1,7 +1,4 @@
-import { getRandomInteger } from './util.js';
-
-const messages = ['Всё отлично!', 'В целом всё неплохо.', 'Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент ?'];
-const names = ['Майра', 'Марс', 'Куся', 'Бэлла', 'Сэм', 'Бэтти', 'Арчи', 'Ричи', 'Чебурашка', 'Пашка'];
+import { getRandomInteger, messages, names } from './utils.js';
 
 const createComment = () => {
   let id = 1;
