+++ js/script.js
@@ -1,1 +1,1 @@
-import './util.js';
+import './utils.js';
